+++ plugins/clickhouse/src/components/panel/LogsDocument.tsx
@@ -8,9 +8,16 @@ import { formatTimeWrapper } from '../../utils/helpers';
 interface ILogsDocumentProps {
   document: IDocument;
   fields?: string[];
+  addFilter?: (filter: string) => void;
+  selectField?: (field: string) => void;
 }
 
-const LogsDocument: React.FunctionComponent<ILogsDocumentProps> = ({ document, fields }: ILogsDocumentProps) => {
+const LogsDocument: React.FunctionComponent<ILogsDocumentProps> = ({
+  document,
+  fields,
+  addFilter,
+  selectField,
+}: ILogsDocumentProps) => {
   const [isExpanded, setIsExpanded] = useState<boolean>(false);
 
   const defaultActions = [
@@ -93,7 +100,7 @@ const LogsDocument: React.FunctionComponent<ILogsDocumentProps> = ({ document, f
       <Tr isExpanded={isExpanded}>
         <Td />
         <Td colSpan={fields && fields.length > 0 ? fields.length + 1 : 2}>
-          {isExpanded && <LogsDocumentDetails document={document} />}
+          {isExpanded && <LogsDocumentDetails document={document} addFilter={addFilter} selectField={selectField} />}
         </Td>
         <Td />
       </Tr>